Introduce annotation processor for plugin ids
The plugin ids are now declared as annotation on the type instead of a
properties file or within the build script. This has several advantages.
To start when searching the code base for the plugin id, we
automatically find the implementation class. We also open the door for
generating an entry class that can be used to give better error message
to the user.