Convert all Kotlin code to Groovy
Both the DSL and plugin Kotlin code was converted to Groovy. For the
life of me, the Kotlin experience in Gradle (both DSL and plugin code)
never ends up to be a smooth ride. Groovy and Java are battle tested
code that just works. I can live without the auto complete inside Groovy
DSL which is just a limitation of the current Groovy DSL implementation.