Use ProjectBuilder to get internal types
Gradle 6.3 changed some internal types which cause the native fixtures
to miss-behave with the latest Gradle. The ProjectBuilder create full
fledge project with a properly configured ServiceRegistry which we can
use to pull the internal type required. This is strictly until we piece
out all the internal type from the native fixtures.