Deprecate gradleFixtures()
The Gradle fixtures was a way to try and extract the conveniences from
Gradle codebase with regards to testing plugins. In the ends, the
conveniences from Gradle codebase contains lots of opinions which ends
up not being easy to use outside of Gradle. We want to approach the
issue differently.

Signed-off-by: Daniel Lacasse <daniel@lacasse.io>