Ensure support for custom plugin source set with various buckets
New source set don't typically create the API, compile only API, API
elements and runtime elements buckets. If this is the case, we will
optionally create the buckets to the best of our ability. Despite the
fact that core Gradle development plugin allow to customize the plugin
source set, the plugin was not coded for that purpose. As leader and
Gradle expert, we write our extension plugins as if it was possible, but
we are most certainly sure they didn't intended to be the case. Users
still have up until after project evaluation to create the missing
buckets.

Signed-off-by: Daniel Lacasse <daniel@lacasse.io>